cameraconnect/requirements.txt
harvesters==1.3.2
numpy==1.19.5
opencv-python==4.5.3.56
paho-mqtt==1.5.1
Pillow==8.3.2
